# specs/backbrain_actions_openapi.yaml
openapi: 3.0.3
info:
  title: Backbrain Actions API
  version: 0.1.0
servers:
  - url: https://CHANGE_ME_TUNNEL_URL
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/summarizer/summarize-file:
    post:
      operationId: summarizeFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, name]
              properties:
                kind: { type: string, enum: [entries, summaries] }
                name: { type: string }
                style: { type: string, enum: [short, bullet, long] }
      responses:
        '200': { description: OK }
  /api/v1/summarizer/summarize-prefix:
    post:
      operationId: summarizePrefix
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, prefix]
              properties:
                kind: { type: string, enum: [entries, summaries] }
                prefix: { type: string }
                limit: { type: integer, default: 50 }
                style: { type: string, enum: [short, bullet, long] }
      responses:
        '200': { description: OK }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key